added default compose config - default compose config from src/main/fabric8-compose - moved kompose handeler to core/services
